refactor(services): restructure service extensions and add metrics service
feat(metrics): implement metrics service interface and implementation
feat(health): add cache and circuit breaker health checks
docs: update readme with comprehensive usage examples
chore: remove unused extension files
