--- src/Acontplus.FactElect/Extensions/ModelExtensions.cs
@@ -1,5 +1,0 @@
-﻿namespace Acontplus.FactElect.Extensions;
-
-public class ModelExtensions
-{
-}
