--- src/Acontplus.FactElect/Extensions/StringExtensions.cs
@@ -1,5 +1,0 @@
-﻿namespace Acontplus.FactElect.Extensions;
-
-public class StringExtensions
-{
-}
