--- src/Acontplus.FactElect/Extensions/XmlExtensions.cs
@@ -1,5 +1,0 @@
-﻿namespace Acontplus.FactElect.Extensions;
-
-public class XmlExtensions
-{
-}
