+++ src/Acontplus.Services/Extensions/ServiceExtensions.cs
@@ -1,4 +1,5 @@
 using Acontplus.Services.Configuration;
+using Acontplus.Services.HealthChecks;
 using Acontplus.Services.Services.Abstractions;
 using Acontplus.Services.Services.Implementations;
 
@@ -215,177 +216,4 @@ public static class ServiceExtensions
     }
 }
 
-/// <summary>
-/// Configuration for caching services.
-/// </summary>
-public class CacheConfiguration
-{
-    /// <summary>
-    /// Use distributed cache (Redis) instead of in-memory cache.
-    /// </summary>
-    public bool UseDistributedCache { get; set; } = false;
 
-    /// <summary>
-    /// Redis connection string.
-    /// </summary>
-    public string? RedisConnectionString { get; set; }
-
-    /// <summary>
-    /// Redis instance name.
-    /// </summary>
-    public string? RedisInstanceName { get; set; }
-
-    /// <summary>
-    /// Memory cache size limit in bytes.
-    /// </summary>
-    public long MemoryCacheSizeLimit { get; set; } = 1024 * 1024 * 100; // 100 MB
-
-    /// <summary>
-    /// Expiration scan frequency in minutes.
-    /// </summary>
-    public int ExpirationScanFrequencyMinutes { get; set; } = 5;
-}
-
-/// <summary>
-/// Interface for metrics service.
-/// </summary>
-public interface IMetricsService
-{
-    void IncrementCounter(string name, Dictionary<string, string>? tags = null);
-    void RecordGauge(string name, double value, Dictionary<string, string>? tags = null);
-    void RecordHistogram(string name, double value, Dictionary<string, string>? tags = null);
-}
-
-/// <summary>
-/// Implementation of metrics service.
-/// </summary>
-public class MetricsService : IMetricsService
-{
-    public void IncrementCounter(string name, Dictionary<string, string>? tags = null)
-    {
-        // Implementation for metrics collection
-    }
-
-    public void RecordGauge(string name, double value, Dictionary<string, string>? tags = null)
-    {
-        // Implementation for metrics collection
-    }
-
-    public void RecordHistogram(string name, double value, Dictionary<string, string>? tags = null)
-    {
-        // Implementation for metrics collection
-    }
-}
-
-/// <summary>
-/// Health check for circuit breaker service.
-/// </summary>
-public class CircuitBreakerHealthCheck : IHealthCheck
-{
-    private readonly ICircuitBreakerService _circuitBreakerService;
-
-    public CircuitBreakerHealthCheck(ICircuitBreakerService circuitBreakerService)
-    {
-        _circuitBreakerService = circuitBreakerService;
-    }
-
-    public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            var defaultState = _circuitBreakerService.GetCircuitBreakerState("default");
-            var apiState = _circuitBreakerService.GetCircuitBreakerState("api");
-            var databaseState = _circuitBreakerService.GetCircuitBreakerState("database");
-            var externalState = _circuitBreakerService.GetCircuitBreakerState("external");
-            var authState = _circuitBreakerService.GetCircuitBreakerState("auth");
-
-            var data = new Dictionary<string, object>
-            {
-                ["default"] = defaultState.ToString(),
-                ["api"] = apiState.ToString(),
-                ["database"] = databaseState.ToString(),
-                ["external"] = externalState.ToString(),
-                ["auth"] = authState.ToString(),
-                ["lastCheckTime"] = DateTime.UtcNow
-            };
-
-            // Check if any critical circuits are open
-            var criticalCircuitsOpen = new[] { databaseState, authState }.Any(state => state == CircuitBreakerState.Open);
-            var anyCircuitOpen = new[] { defaultState, apiState, databaseState, externalState, authState }
-                .Any(state => state == CircuitBreakerState.Open);
-
-            if (criticalCircuitsOpen)
-            {
-                return Task.FromResult(HealthCheckResult.Unhealthy("Critical circuit breakers are open", data: data));
-            }
-
-            if (anyCircuitOpen)
-            {
-                return Task.FromResult(HealthCheckResult.Degraded("Some circuit breakers are open", data: data));
-            }
-
-            return Task.FromResult(HealthCheckResult.Healthy("All circuit breakers are operational", data));
-        }
-        catch (Exception ex)
-        {
-            return Task.FromResult(HealthCheckResult.Unhealthy("Circuit breaker service failed", ex));
-        }
-    }
-}
-
-/// <summary>
-/// Health check for cache service.
-/// </summary>
-public class CacheHealthCheck : IHealthCheck
-{
-    private readonly ICacheService _cacheService;
-
-    public CacheHealthCheck(ICacheService cacheService)
-    {
-        _cacheService = cacheService;
-    }
-
-    public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            // Test cache functionality with a comprehensive test
-            var testKey = $"health-check-{Guid.NewGuid():N}";
-            var testValue = $"test-{DateTime.UtcNow:yyyy-MM-dd-HH-mm-ss}";
-            var testExpiration = TimeSpan.FromSeconds(30);
-
-            // Test Set operation
-            await _cacheService.SetAsync(testKey, testValue, testExpiration, cancellationToken);
-
-            // Test Get operation
-            var retrieved = await _cacheService.GetAsync<string>(testKey, cancellationToken);
-
-            // Test Exists operation
-            var exists = await _cacheService.ExistsAsync(testKey, cancellationToken);
-
-            // Test Remove operation
-            await _cacheService.RemoveAsync(testKey, cancellationToken);
-
-            // Verify removal
-            var removedValue = await _cacheService.GetAsync<string>(testKey, cancellationToken);
-
-            if (retrieved == testValue && exists && removedValue == null)
-            {
-                var stats = _cacheService.GetStatistics();
-                var data = new Dictionary<string, object>
-                {
-                    ["totalEntries"] = stats.TotalEntries,
-                    ["hitRatePercentage"] = stats.HitRatePercentage,
-                    ["lastTestTime"] = DateTime.UtcNow
-                };
-                return HealthCheckResult.Healthy("Cache service is fully operational", data);
-            }
-
-            return HealthCheckResult.Degraded("Cache service test partially failed - some operations may not be working correctly");
-        }
-        catch (Exception ex)
-        {
-            return HealthCheckResult.Unhealthy("Cache service failed", ex);
-        }
-    }
-}
