+++ src/Acontplus.Services/Services/Implementations/MetricsService.cs
@@ -1,0 +1,60 @@
+using Acontplus.Services.Services.Abstractions;
+
+namespace Acontplus.Services.Services.Implementations;
+
+/// <summary>
+/// Implementation of metrics service.
+/// </summary>
+public class MetricsService : IMetricsService
+{
+    private readonly ILogger<MetricsService> _logger;
+
+    public MetricsService(ILogger<MetricsService> logger)
+    {
+        _logger = logger;
+    }
+
+    /// <summary>
+    /// Increment a counter metric.
+    /// </summary>
+    /// <param name="name">The name of the counter.</param>
+    /// <param name="tags">Optional tags for the metric.</param>
+    public void IncrementCounter(string name, Dictionary<string, string>? tags = null)
+    {
+        // Implementation for metrics collection
+        _logger.LogDebug("Incrementing counter: {CounterName} with tags: {@Tags}", name, tags);
+
+        // TODO: Implement actual metrics collection (e.g., Application Insights, Prometheus, etc.)
+        // For now, this is a placeholder implementation
+    }
+
+    /// <summary>
+    /// Record a gauge metric.
+    /// </summary>
+    /// <param name="name">The name of the gauge.</param>
+    /// <param name="value">The value to record.</param>
+    /// <param name="tags">Optional tags for the metric.</param>
+    public void RecordGauge(string name, double value, Dictionary<string, string>? tags = null)
+    {
+        // Implementation for metrics collection
+        _logger.LogDebug("Recording gauge: {GaugeName} = {Value} with tags: {@Tags}", name, value, tags);
+
+        // TODO: Implement actual metrics collection (e.g., Application Insights, Prometheus, etc.)
+        // For now, this is a placeholder implementation
+    }
+
+    /// <summary>
+    /// Record a histogram metric.
+    /// </summary>
+    /// <param name="name">The name of the histogram.</param>
+    /// <param name="value">The value to record.</param>
+    /// <param name="tags">Optional tags for the metric.</param>
+    public void RecordHistogram(string name, double value, Dictionary<string, string>? tags = null)
+    {
+        // Implementation for metrics collection
+        _logger.LogDebug("Recording histogram: {HistogramName} = {Value} with tags: {@Tags}", name, value, tags);
+
+        // TODO: Implement actual metrics collection (e.g., Application Insights, Prometheus, etc.)
+        // For now, this is a placeholder implementation
+    }
+}
